+++ imports/api/chats/chat.publications.js
@@ -1,0 +1,13 @@
+import { Meteor } from 'meteor/meteor';
+import { getHost } from '../@/shared';
+import Chats from './chat';
+
+Meteor.publish('chat', function (contextId) {
+  const host = getHost(this);
+  const user = Meteor.user();
+  if (user) {
+    return Chats.find({
+      contextId: contextId,
+    });
+  }
+});
