+++ imports/ui/RouterComponents/processes/ProcessContainer.js
@@ -1,6 +1,7 @@
 import { withTracker } from 'meteor/react-meteor-data';
 import Process from './Process';
 import Processes from '../../../api/processes/process';
+import Chats from '../../../api/chats/chat';
 
 export default ProcessContainer = withTracker((props) => {
   const processId = props.match.params.id;
