+++ lib/collections.js
@@ -8,6 +8,6 @@ Activities = new Mongo.Collection('activities');
 // Pages = new Mongo.Collection('pages');
 
 Resources = new Mongo.Collection('resources');
-Chats = new Mongo.Collection('chats');
+// Chats = new Mongo.Collection('chats');
 // Documents = new Mongo.Collection('documents');
 // Categories = new Mongo.Collection('categories');
